Grocery Application
You would be surprised to know that the general public has recognized the absolute necessity of these services during the COVID-19 lockdown phase. In this part, the grocery app plays important in day-to-day life for buying essential items like Veggies, milk packets, fruits, and many more.

Our app does it, but with a twisted approach.Our app holds the KISS principle which helps our app to reach the benchmark of simplicity and easy to use. In this Application works with CRUD operation. It's lightweight in nature and it can operate on 96% of mobile systems.

ScreenShot:-